Make DSL engine initialization more robust.
The DSL engine can now be re-initializes by simply calling the
`dslengine` package `Reset` function. Plugins register their DSL roots
with the `Register` package function. The roots are available via the
`Roots` package function.
