+++ design/apidsl/media_type_test.go
@@ -15,7 +15,7 @@ var _ = Describe("MediaType", func() {
 	var mt *MediaTypeDefinition
 
 	BeforeEach(func() {
-		InitDesign()
+		Reset()
 		dslengine.Errors = nil
 		name = ""
 		dslFunc = nil
@@ -211,7 +211,7 @@ var _ = Describe("Duplicate media types", func() {
 	}
 
 	BeforeEach(func() {
-		InitDesign()
+		Reset()
 		dslengine.Errors = nil
 		mt = MediaType(id, dslFunc)
 		duplicate = MediaType(id, dslFunc)
@@ -244,7 +244,7 @@ var _ = Describe("CollectionOf", func() {
 	Context("used on a global variable", func() {
 		var col *MediaTypeDefinition
 		BeforeEach(func() {
-			InitDesign()
+			Reset()
 			mt := MediaType("application/vnd.example", func() {
 				Attribute("id")
 				View("default", func() {
@@ -272,7 +272,7 @@ var _ = Describe("CollectionOf", func() {
 	Context("defined with the media type identifier", func() {
 		var col *MediaTypeDefinition
 		BeforeEach(func() {
-			InitDesign()
+			Reset()
 			MediaType("application/vnd.example+json", func() {
 				Attribute("id")
 				View("default", func() {
@@ -314,7 +314,7 @@ var _ = Describe("CollectionOf", func() {
 var _ = Describe("Example", func() {
 	Context("defined examples in a media type", func() {
 		BeforeEach(func() {
-			InitDesign()
+			Reset()
 		})
 
 		It("produces a media type with examples", func() {
