+++ design/apidsl/type_test.go
@@ -15,7 +15,7 @@ var _ = Describe("Type", func() {
 	var ut *UserTypeDefinition
 
 	BeforeEach(func() {
-		InitDesign()
+		Reset()
 		dslengine.Errors = nil
 		name = ""
 		dsl = nil
