+++ goagen/gen_swagger/swagger_test.go
@@ -30,7 +30,7 @@ var _ = Describe("New", func() {
 	BeforeEach(func() {
 		swagger = nil
 		newErr = nil
-		InitDesign()
+		Reset()
 		genschema.Definitions = make(map[string]*genschema.JSONSchema)
 	})
