+++ goagen/meta/generator_test.go
@@ -256,7 +256,7 @@ invalid go code
 
 	panickySource = `package foo
 import "github.com/goadesign/goa/dslengine"
-func Generate(roots dslengine.RootDefinitions) ([]string, error) {
+func Generate(roots []dslengine.Root) ([]string, error) {
 	return nil, nil
 }
 
@@ -265,7 +265,7 @@ func init() { panic("kaboom") }
 
 	validSource = `package foo
 import "github.com/goadesign/goa/dslengine"
-func Generate(roots dslengine.RootDefinitions) ([]string, error) {
+func Generate(roots []dslengine.Root) ([]string, error) {
 	return nil, nil
 }
 `
@@ -273,7 +273,7 @@ func Generate(roots dslengine.RootDefinitions) ([]string, error) {
 	validSourceTmpl = `package foo
 import "fmt"
 import "github.com/goadesign/goa/dslengine"
-func Generate(roots dslengine.RootDefinitions) ([]string, error) {
+func Generate(roots []dslengine.Root) ([]string, error) {
 	{{range .}}fmt.Println("{{.}}")
 	{{end}}
 	return nil, nil
